refactor: 마이그레이션으로 사용 안되는 기존 파일 제거 및 패키지 이동
