+++ android/app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".ui.setting.SettingActivity"
+            android:name=".ui.mypage.MyPageActivity"
             android:exported="false" />
         <activity
             android:name=".ui.login.LoginActivity"
@@ -49,9 +49,6 @@
             android:name=".ui.main.MainActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
-        <activity
-            android:name=".ui.mypage.MyPageActivity"
-            android:exported="false" />
 
         <meta-data
             android:name="com.google.android.geo.API_KEY"
