+++ android/app/src/main/java/com/on/turip/ui/compose/mypage/model/AppEnvironmentInfoModel.kt
@@ -1,4 +1,4 @@
-package com.on.turip.ui.compose.setting.model
+package com.on.turip.ui.compose.mypage.model
 
 data class AppEnvironmentInfoModel(
     val appVersionName: String,
