--- android/app/src/main/java/com/on/turip/ui/compose/setting/SettingUiEffect.kt
@@ -1,17 +1,0 @@
-package com.on.turip.ui.compose.setting
-
-import com.on.turip.ui.common.error.ErrorUiState
-
-sealed interface SettingUiEffect {
-    data object NavigateToLogin : SettingUiEffect
-
-    data class ShowError(
-        val errorUiState: ErrorUiState,
-        val retryAction: SettingRetryAction,
-    ) : SettingUiEffect
-}
-
-enum class SettingRetryAction {
-    LOGOUT,
-    WITHDRAW,
-}
