--- android/app/src/main/java/com/on/turip/ui/compose/setting/component/SettingItem.kt
@@ -1,62 +1,0 @@
-package com.on.turip.ui.compose.setting.component
-
-import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.on.turip.R
-import com.on.turip.ui.compose.designsystem.theme.TuripTheme
-
-@Composable
-fun SettingItem(
-    @DrawableRes imageRes: Int,
-    @StringRes titleRes: Int,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier =
-            modifier
-                .clickable(onClick = onClick)
-                .padding(horizontal = TuripTheme.spacing.extraLarge, vertical = TuripTheme.spacing.large),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Image(
-            painter = painterResource(imageRes),
-            contentDescription = null,
-            modifier =
-                Modifier
-                    .padding(end = TuripTheme.spacing.large)
-                    .size(24.dp),
-        )
-        Text(
-            text = stringResource(titleRes),
-            style = TuripTheme.typography.body1,
-            textAlign = TextAlign.Center,
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun SettingItemPreview() {
-    TuripTheme {
-        SettingItem(
-            imageRes = R.drawable.ic_inquire,
-            titleRes = R.string.setting_inquiry,
-            onClick = {},
-        )
-    }
-}
